2024 12 update (#7)
* RabbitMQ on amd64/arm64; Ubumtu 24.04 Dec 2024; pre-Nitro volume mounting

* Versions in the config, not in sh

* Comments

* Added IAM screenshots to readme

---------

+++ pkg/rexec/scripts/nginx/install.sh
@@ -1,6 +1,11 @@
 # https://www.digitalocean.com/community/tutorials/how-to-configure-nginx-as-a-reverse-proxy-on-ubuntu-22-04
 
-sudo DEBIAN_FRONTEND=noninteractive apt-get -y install nginx
+# apt-get install has a habit to write "Running kernel seems to be up-to-date." to stderr. Ignore it and rely on the exit code
+sudo DEBIAN_FRONTEND=noninteractive apt-get install -y nginx 2>/dev/null
+if [ "$?" -ne "0" ]; then
+    echo nginx install error, exiting
+    exit $?
+fi
 
 # Remove nginx stub site
 sudo rm -f /etc/nginx/sites-enabled/default
